# Backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_odatlar-bot_backend_internal_dto.CreatePracticeRequest:
    properties:
      benefits:
        items:
          type: string
        maxItems: 20
        type: array
      name:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  github_com_odatlar-bot_backend_internal_dto.PracticeHistoryEntryResponse:
    properties:
      completedAt:
        type: string
      date:
        type: string
      habitId:
        type: string
      habitName:
        type: string
      id:
        type: string
      kind:
        type: string
    type: object
  github_com_odatlar-bot_backend_internal_dto.PracticeHistoryResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.PracticeHistoryEntryResponse'
        type: array
    type: object
  github_com_odatlar-bot_backend_internal_dto.PracticeListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.PracticeResponse'
        type: array
    type: object
  github_com_odatlar-bot_backend_internal_dto.PracticeResponse:
    properties:
      benefits:
        items:
          type: string
        type: array
      completedToday:
        type: boolean
      createdAt:
        type: string
      id:
        type: string
      kind:
        type: string
      level:
        type: integer
      name:
        type: string
      streak:
        type: integer
      xp:
        type: integer
      xp_reward:
        type: integer
    type: object
...
  github_com_odatlar-bot_backend_internal_dto.UpdatePracticeRequest:
    properties:
      benefits:
        items:
          type: string
        maxItems: 20
        type: array
      name:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  github_com_odatlar-bot_backend_pkg_response.ErrorBody:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@odatlar-bot.uz
    name: API Support
  description: Admin panel, Bot Runtime (Telegram WebApp), Qo'llanma (videolar, kurslar,
    fayllar) va Bildirishnomalar APIlari.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Odatlar Bot API
  version: "1.0"
paths:
...
  /bot-runtime/practices:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.PracticeListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - TelegramInitData: []
      summary: Amaliyotlar ro'yxati
      tags:
      - Practices
    post:
      consumes:
      - application/json
      parameters:
      - description: Amaliyot ma'lumotlari
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.CreatePracticeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.PracticeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - TelegramInitData: []
      summary: Yangi amaliyot yaratish
      tags:
      - Practices
  /bot-runtime/practices/{id}:
    delete:
      parameters:
      - description: Amaliyot ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - TelegramInitData: []
      summary: Amaliyotni o'chirish
      tags:
      - Practices
    put:
      consumes:
      - application/json
      parameters:
      - description: Amaliyot ID
        in: path
        name: id
        required: true
        type: integer
      - description: Yangilangan ma'lumotlar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.UpdatePracticeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.PracticeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - TelegramInitData: []
      summary: Amaliyotni yangilash
      tags:
      - Practices
  /bot-runtime/practices/{id}/toggle:
    post:
      parameters:
      - description: Amaliyot ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.PracticeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - TelegramInitData: []
      summary: Amaliyotni bajarilgan/bajarilmagan qilish
      tags:
      - Practices
  /bot-runtime/practices/history:
    get:
      parameters:
      - description: Boshlanish sanasi (YYYY-MM-DD)
        in: query
        name: from
        required: true
        type: string
      - description: Tugash sanasi (YYYY-MM-DD)
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.PracticeHistoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - TelegramInitData: []
      summary: Amaliyotlar tarixi
      tags:
      - Practices
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token. Format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
  BotToken:
    description: 'Telegram bot token. Format: Bearer {bot_token}'
    in: header
    name: Authorization
    type: apiKey
  TelegramInitData:
    description: 'Telegram WebApp initData. Alternativa: Authorization: tma {initData}'
    in: header
    name: X-Telegram-Init-Data
    type: apiKey
swagger: "2.0"